1. Using three environment variables: - APPOPTICS_SERVICE_KEY has to be in the right format. No connection is actually made with it. - AO_TEST_COLLECTOR a url of a collector to use - AO_TEST_SERVICE_KEY a valid service key for the collector above. 2. token test is now part of core tests. 3. separated connection test (only one not using udp) to own test.
